Fixed sender script

+++ sender/remindme.sh
@@ -1,10 +1,24 @@
 #!/bin/bash
-timestamp=$(date +%s)
-fname="saying_$timestamp.txt"
 HOST=yapper
-echo $fname
-echo $1 "Current time is $(date +'%H:%M') Reminder for " > fname
-echo $1 >>  $fname
-echo $1
-#scp  $fname pi@$HOST:~/talk/sayings/
-scp  $fname $HOST:~/talk/sayings/
+dir=/home/avi/reminder
+sayingdir=$dir/sayings
+
+if [ ! -d $sayingdir ]; then
+	mkdir $sayingdir
+fi
+
+timestamp=$(date +%s)
+t=` echo $1 | /usr/bin/md5sum  | awk '{ print $1 }' `
+fname="${t}_$(date +%F).txt"
+fname=$sayingdir/$fname
+if [ ! -f $fname ]; then
+	echo $fname $1
+	echo "Current time is $(date +'%I:%M') Reminder for " > $fname
+	echo $1 >>  $fname
+	#scp  $fname pi@$HOST:~/talk/sayings/
+	scp  $fname $HOST:~/talk/sayings/
+	cat /dev/null > $fname
+else
+	echo "File exists already $fname " 
+
+fi
